refactor: remove overengineered tokenusage package and SQLite dependency
删除 internal/tokenusage/ 整个包（约1200行），这是一个无消费者的过度设计：

- TokenUsageManager 日/周/月聚合、TPS/延迟/错误率统计、告警阈值检查、
  CSV/JSON 导出等全部功能无 TUI 或 CLI 消费者
- SQLite 存储路径永远不可达（go.mod 无 driver 依赖，静默降级到 memory）
- TUI 状态栏 token 显示直接使用 LLM 响应中的 llm.Usage 字段，
  不经过 TokenUsageManager
- recordTokenUsage() 写入数据后无人读取，GetTokenRealtimeSnapshot() 仅被测试调用

保留：
- tui/token_monitor.go 的 tokenUsageComponent（TUI 状态栏 token 显示）
- emitUsageEvent() 事件发射逻辑（移至 runner.go）
- ApproximateTokens/ApproximateRequestTokens 移至 internal/context/builder.go

同时更新 docs/architecture.md，新增完整用户故事走查和12个关键设计决策说明。
